pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "odoo-launch"
version = "0.1.0"
description = "Disposable Odoo launcher for automated testing."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Loomworks Automation" }
]
dependencies = [
  "typer[all]>=0.12.3",
  "pyyaml>=6.0.1",
  "jinja2>=3.1.2",
  "rich>=13.7.1",
  "httpx>=0.27.0",
  "psutil>=5.9.0",
  "psycopg[binary]>=3.1.18"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-mock>=3.12.0",
  "pytest-cov>=5.1.0",
  "ruff>=0.8.0",
  "mypy>=1.13.0",
  "types-PyYAML>=6.0.12.20241230"
]

[project.scripts]
odoo-launch = "cli.main:app"

[tool.setuptools]
packages = ["cli", "scripts"]
include-package-data = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "--maxfail=1 --disable-warnings -q"
testpaths = [
  "tests"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "W", "I", "C90"]
ignore = ["E501", "B008"]
